Add package.json
Added package.json. I don't really know what package.json is for.
I'm a noob. But I know that it's necessary for this method of bot
creation.

I used a previous package.json as a template for this one. The MS Bot
Framework has annoying intricacies. It is picky with package.json
files. Many of my package.json files caused program crashes.

My last package.json was the only working one I had. I have no idea what
made it work. But because it worked, I used it as a template for this
package.json.

I simply changed the "name", "description", "url", etc. as necessary.